[BOTF-351] Payment Method (#370)
* [BOTF-352] Installing Stripe (#327)
* [BOTF-353] Home Plans Section (#328)
* [BOTF-354] Admin Painel Foundation (#329)
* [BOTF-355] Listing Plans on Home Page (#330)
* [BOTF-356] Subscribe Plans Page (#331)
* [BOTF-357] User Address Fields (#332)
* [BOTF-358] Create Subscription (#333)
* [BOTF-359] Redirections for subscribe (#334)
* [BOTF-360] Refactoring Authentication (#335)
* [BOTF-351] Adding rule to Admin page
* [BOTF-361] Listing Stripe Data (#336)
* [BOTF-362] Create User on Admin side (#337)
* [BOTF-363] Admin edit user and Profile Page (#338)
* [BOTF-364] Admin Top Menu (#339)
* [BOTF-365] Improving UX/UI of Plans (#340)
* [BOTF-366] Upgrading Subscriptions (#341)
* [BOTF-351] Fixing PlanCard issue on prices
* [BOTF-367] Listing User subscription on Profile (#342)
* [BOTF-368] Listing User Transactions (#343)
* [BOTF-369] Issues listing transactions (#344)
* [BOTF-370] Issues populating user (#345)
* [BOTF-371] Improving Checkout Workflow (#346)
* [BOTF-374] Improve Subscribe Plan Page (#347)
* [BOTF-373] Fill Required Address (#348)
* [BOTF-372] Header Icons (#349)
* [BOTF-377] Canceling User Subscription (#350)
* [BOTF-376] Apply Coupons (#351)
* [BOTF-375] Mobile Issues for new pages (#352)
* [BOTF-382] Display Coupons Discounts (#353)
* [BOTF-381] Pricing Page (#354)
* [BOTF-385] How It Works Foundation (#355)
* [BOTF-379] Date Picker
* [BOTF-378] Docker Tests - Small Issues (#356)
* [BOTF-383] Upload Feature (#357)
* [BOTF-386] Implementing Avatars (#358)
* [BOTF-387] Docker Tests (#359)
* [BOTF-388] How it works content (#360)
* [BOTF-380] Google Analytics (#361)
* [BOTF-389] Issues updating avatar on the screen (#362)
* [BOTF-392] Change Bot selection to Search select input (#363)
* [BOTF-393] Prevent Google Bot on Auth areas
* [BOTF-394] Switching to Search box the indicators picker (#364)
* [BOTF-395] Copy strategy button (#365)
* [BOTF-396] Import and Export Pilots (#366)
* [BOTF-351] Update Binance-Sync
* [BOTF-398] Limit Wallets and Slots according to the Plan (#367)
* [BOTF-400] Replace "Master Account" for "Wallet" (#368)
* [BOTF-399] Fix Mobile Issues (#369)

+++ src/contexts/AuthUser.jsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { createContext, useEffect, useState, useContext, useRef } from 'react';
-import { useRouter } from 'next/navigation';
+import { useRouter, useSearchParams } from 'next/navigation';
 import APIContext from './4HandsAPI';
 import LoadingPage from '@/app/loading';
 import ErrorPage from '@/app/error';
@@ -23,34 +23,70 @@ export default AuthUserContext;
  *
  * @returns {JSX.Element} The context provider component with loading and error handling.
  */
-export function AuthUserProvider({ children, ...props }) {
-   const [ userAuth, setUserAuth ] = useState();
-   const [ error, setError ] = useState();
+export function AuthUserProvider({ children, rules = [], ...props }) {
+   const [userAuth, setUserAuth] = useState();
+   const [error, setError] = useState();
    const instance = useContext(APIContext);
    const router = useRouter();
    const userChecked = useRef();
+   const searchParams = useSearchParams();
+   const updatePlan = searchParams.get('updatePlan');
+   const isUpdatePlan = updatePlan === 'true';
 
    useEffect(() => {
       if (userChecked.current) return;
       userChecked.current = true;
 
       instance.auth.checkUser().then(authData => {
+         const url = new URL(window.location);
+         const pathname = window.location.pathname;
+
          if (authData.error) {
             return setError(authData);
          }
 
          if (authData.isLogged) {
-            const userLetters = authData?.user?.fullName.split(' ').map(word => word[0]?.toUpperCase() || '').join('');
+            const userRoles = authData?.user?.rules || [];
+            const hasAddress = !!authData?.user?.billingAddress;
+            const hasRequiredRole = rules.length === 0 || rules.some(role => userRoles.includes(role));
+
+            if (!hasRequiredRole) {
+               return setError({ error: true, message: 'Unauthorized: Insufficient permissions' });
+            }
+
+            if (!hasAddress && window.location.pathname !== '/dashboard/user/my-profile') {
+               return router.push('/dashboard/user/my-profile');
+            }
+
+            if (authData.name === 'USER_EMAIL_NOT_CONFIRMED') {
+               return setError(authData);
+            }
+
+            const userLetters = authData?.user?.fullName?.split(' ').map(word => word[0]?.toUpperCase() || '').join('') || '';
 
             window.localStorage.setItem('userLetters', userLetters);
             setUserAuth(authData);
+
+            if ((pathname !== '/subscribe-plan' && pathname !== '/admin') && !authData?.user?.subscribedPlan) {
+               return router.push('/subscribe-plan');
+            } else if (pathname === '/subscribe-plan' && authData?.user?.subscribedPlan && !isUpdatePlan) {
+               return router.push('/dashboard');
+            }
          } else {
-            router.push('/dashboard/login');
+            url.pathname = '/dashboard/login';
+
+            if (pathname === '/subscribe-plan') {
+               url.searchParams.set('register', 'true');
+            } else {
+               url.searchParams.set('redirect', pathname);
+            }
+
+            router.push(url.toString());
          }
       }).catch(err => {
          setError(err);
       });
-   }, [ instance.auth, router ]);
+   }, [instance.auth, router]);
 
    return (
       <AuthUserContext.Provider value={userAuth}>
